requirements.txt
requests==2.31.0
pandas==2.1.4
geopandas==0.14.1
folium==0.15.1
matplotlib==3.8.2
seaborn==0.13.0
numpy>=1.26
python-dotenv==1.0.0
cachetools==5.3.2
shapely==2.0.2
scipy
